local-retail-platform/backend/pyproject.toml
[project]
name = "local-retail-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Tomy Huy Le",email = "tole02@dfki.de"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-filter (>=25.2,<26.0)",
    "pydantic (>=2.12.0,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "django-stubs (>=5.2.7,<6.0.0)",
    "djangorestframework-stubs (>=3.16.4,<4.0.0)"
]
